* Add change to characters.sql from last commit
--HG--
branch : trunk

+++ sql/characters.sql
@@ -375,7 +375,7 @@ CREATE TABLE `characters` (
   `yesterdayKills` smallint(5) unsigned NOT NULL default'0',
   `chosenTitle` int(10) unsigned NOT NULL default'0',
   `knownCurrencies` bigint(20) unsigned NOT NULL default'0',
-  `watchedFaction` int(10) unsigned NOT NULL default'0',
+  `watchedFaction` bigint(10) unsigned NOT NULL default'0',
   `drunk` smallint(5) unsigned NOT NULL default'0',
   `health` int(10)  unsigned NOT NULL default'0',
   `power1` int(10) unsigned NOT NULL default'0',
